pyproject.toml
[project]
name = "codevul"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.2.0",
    "httpx[socks]>=0.28.1",
    "ijson>=3.4.0.post0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "tree-sitter>=0.25.2",
    "tree-sitter-c>=0.24.1",
    "tree-sitter-cpp>=0.23.4",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-languages>=1.10.2",
]
